requirements.txt
Flask
python-dotenv
jira==3.8.0